+++ crates/ludtwig/src/check/rules.rs
@@ -4,6 +4,7 @@ use crate::check::rules::html_string_quotation::RuleHtmlStringQuotation;
 use crate::check::rules::indentation::RuleIndentation;
 use crate::check::rules::line_ending::RuleLineEnding;
 use crate::check::rules::ludtwig_ignore_file_not_on_top::RuleLudtwigIgnoreFileNotOnTop;
+use crate::check::rules::twig_block_duplicate::RuleTwigBlockDuplicate;
 use crate::check::rules::twig_block_line_breaks::RuleTwigBlockLineBreaks;
 use crate::check::rules::twig_block_name_snake_case::RuleTwigBlockNameSnakeCase;
 use crate::check::rules::twig_hash_key_no_quotes::RuleTwigHashKeyNoQuotes;
@@ -24,6 +25,7 @@ mod html_string_quotation;
 mod indentation;
 mod line_ending;
 mod ludtwig_ignore_file_not_on_top;
+mod twig_block_duplicate;
 mod twig_block_line_breaks;
 mod twig_block_name_snake_case;
 mod twig_hash_key_no_quotes;
@@ -52,6 +54,7 @@ pub static RULE_DEFINITIONS: &[&'static dyn Rule] = &[
     &RuleTwigPreferShopwareExtends,
     &RuleTwigUseIsSameAs,
     &RuleTwigUseIsNotSameAs,
+    &RuleTwigBlockDuplicate,
 ];
 
 /// Get active rule definitions based on config