+++ crates/ludtwig/CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#102](https://github.com/MalteJanz/ludtwig/issues/102) [BREAKING] Added `version` to `ludtwig-config.toml`, which
   should match the used ludtwig version. Ludtwig will display a warning if this doesn't match, to ensure the user keeps
   their config up to date and is aware of the latest feature additions (doesn't miss out on rules for example).
+- [#50](https://github.com/MalteJanz/ludtwig/pull/50) Added rule `twig-block-duplicate`
 
 # v0.8.3